Remove unused PFX work, delete cert modal, refactor turnonmdmmessage component for reuse
