+++ frontend/components/FileUploader/FileUploader.tsx
@@ -18,7 +18,6 @@ export type ISupportedGraphicNames = Extract<
   | "file-py"
   | "file-script"
   | "file-pdf"
-  | "file-pfx"
   | "file-pkg"
   | "file-p7m"
   | "file-pem"
