--- frontend/components/graphics/FilePfx.tsx
@@ -1,57 +1,0 @@
-import React from "react";
-
-const FilePfx = () => {
-  return (
-    <svg xmlns="http://www.w3.org/2000/svg" width="34" height="40" fill="none">
-      <g clipPath="url(#file-pfx)">
-        <path
-          d="M29.333 39.75H4.667a2.417 2.417 0 0 1-2.417-2.416V2.667A2.417 2.417 0 0 1 4.667.25h19.562c.64 0 1.255.255 1.709.708l5.104 5.105c.453.453.708 1.068.708 1.709v29.562a2.417 2.417 0 0 1-2.417 2.416Z"
-          fill="#fff"
-          stroke="#192147"
-          strokeWidth=".5"
-        />
-        <path
-          d="M23.5.5h.834l.5 6.5 6.666.5v1h-6a2 2 0 0 1-2-2v-6Z"
-          fill="#C5C7D1"
-        />
-        <path
-          d="M24.5.334v5.667c0 .736.597 1.333 1.333 1.333h6"
-          stroke="#192147"
-          strokeWidth=".5"
-        />
-        <path
-          d="M2.5 20h25a2 2 0 0 1 2 2v13a2 2 0 0 1-2 2h-25V20Z"
-          fill="#C5C7D1"
-        />
-        <rect
-          x=".25"
-          y="18.25"
-          width="27.7"
-          height="16.35"
-          rx="1.75"
-          fill="#515774"
-        />
-        <rect
-          x=".25"
-          y="18.25"
-          width="27.7"
-          height="16.35"
-          rx="1.75"
-          stroke="#192147"
-          strokeWidth=".5"
-        />
-        <path
-          d="m16.188 24.16.796 1.69.838-1.69h1.946l-1.463 2.726 1.534 2.728h-1.932l-.923-1.733-.895 1.733h-1.96l1.548-2.728-1.449-2.727h1.96ZM13.742 24.16v1.42h-3.65v-1.42h3.65Zm-2.955 5.454v-5.611c0-.45.08-.823.242-1.119.16-.296.388-.517.681-.664.294-.147.64-.22 1.037-.22.249 0 .49.019.725.057.237.038.412.07.525.1l-.284 1.405a1.531 1.531 0 0 0-.248-.05 1.703 1.703 0 0 0-.249-.02c-.185 0-.309.038-.373.117a.473.473 0 0 0-.096.309v5.696h-1.96ZM4 31.66v-7.5h1.946v.95h.043c.07-.184.175-.353.312-.507.137-.156.308-.28.511-.373.204-.095.44-.142.71-.142.36 0 .703.096 1.027.288.327.191.592.493.795.905.206.412.31.947.31 1.605 0 .63-.099 1.152-.295 1.566-.194.415-.455.724-.782.927a1.97 1.97 0 0 1-1.068.306c-.256 0-.485-.042-.686-.125-.199-.085-.37-.2-.511-.344a1.593 1.593 0 0 1-.323-.497H5.96v2.94H4Zm1.918-4.774c0 .265.034.495.103.69.07.191.17.34.298.447.13.104.285.156.465.156.18 0 .333-.05.458-.153a.96.96 0 0 0 .291-.444c.069-.194.103-.426.103-.696 0-.27-.034-.5-.103-.692a.936.936 0 0 0-.29-.444.693.693 0 0 0-.459-.156.722.722 0 0 0-.465.156.981.981 0 0 0-.298.444 2.062 2.062 0 0 0-.103.692Z"
-          fill="#fff"
-        />
-      </g>
-      <defs>
-        <clipPath id="file-pfx">
-          <path fill="#fff" d="M0 0h34v40H0z" />
-        </clipPath>
-      </defs>
-    </svg>
-  );
-};
-
-export default FilePfx;
