+++ frontend/pages/admin/IntegrationsPage/cards/MdmSettings/AppleBusinessManagerPage/_styles.scss
@@ -22,8 +22,7 @@
     }
   }
 
-
-  &__turn-on-mdm-message, &__add-adm-message {
+  &__add-adm-message {
     margin: 0 auto;
     text-align: center;
     width: 450px;
