+++ frontend/pages/admin/IntegrationsPage/cards/MdmSettings/AppleMdmPage/_styles.scss
@@ -35,7 +35,7 @@
   &__setup-instructions-list {
     padding: 0;
     margin: 0;
-    // list-style: none;
+    list-style: none;
     display: flex;
     flex-direction: column;
     gap: $pad-large;
