--- frontend/pages/admin/IntegrationsPage/cards/MdmSettings/ScepPage/DeleteScepModal/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./DeleteScepModal";
